only cast null constructor args for primitives (#520)
* only cast null for primitives

before, we tried to cast applied types to the type arguments, to capture context bounds. That fails for simple generic types
that then get cast to the wrong type. However, that's all unnessecary: the only nulls we need to cast are primitives, which can't
be passed as null. Anyting else doesn't need a cast at all

* add failing tests

* ignore failing tests

* fix class type parameters

* add tests from passer-by

* move test that don't (yet) pass on scala 2 to scala 3 specific testsuite

* simply getting target